# document-engine/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/v1/fields/{id}": {
      "patch": {
        "tags": [
          "field-correction-controller"
        ],
        "operationId": "correct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DecisionView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "type": "string"
          },
          "previousValue": {
            "type": "object"
          },
          "newValue": {
            "type": "object"
          },
          "reason": {
            "type": "string"
          },
          "decidedBy": {
            "type": "string",
            "format": "uuid"
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CorrectionRequest": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "pageIndex": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Result": {
        "type": "object",
        "properties": {
          "fieldId": {
            "type": "string",
            "format": "uuid"
          },
          "fieldName": {
            "type": "string"
          },
          "machineValue": {
            "type": "object"
          },
          "effectiveValue": {
            "type": "object"
          },
          "reviewStatus": {
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/DecisionView"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}